schema/generateDomainClasses: make lazy using Def.taskDyn
only recompile generated sources if necessary: it's taking a long time and emits lots of warnings

+++ schema/build.sbt
@@ -4,9 +4,21 @@ libraryDependencies += "io.shiftleft" %% "overflowdb-codegen" % "4af837793cf85d2
 
 val generateDomainClasses = taskKey[Seq[File]]("generate overflowdb domain classes for our schema")
 
-generateDomainClasses := {
+generateDomainClasses := Def.taskDyn {
   val outputRoot = target.value / "odb-codegen"
-  FileUtils.deleteRecursively(outputRoot)
-  val invoked = (Compile/runMain).toTask(s" io.shiftleft.codepropertygraph.schema.Codegen schema/target/odb-codegen").value
-  FileUtils.listFilesRecursively(outputRoot)
-}
+  val currentMd5 = FileUtils.md5(sourceDirectory.value)
+
+  if (!outputRoot.exists || CodeGenGlobalState.lastMd5 != currentMd5) {
+    Def.task {
+      FileUtils.deleteRecursively(outputRoot)
+      val invoked = (Compile/runMain).toTask(s" io.shiftleft.codepropertygraph.schema.Codegen schema/target/odb-codegen").value
+      CodeGenGlobalState.lastMd5 = currentMd5
+      FileUtils.listFilesRecursively(outputRoot)
+    }
+  } else {
+    Def.task {
+      CodeGenGlobalState.lastMd5 = currentMd5
+      FileUtils.listFilesRecursively(outputRoot)
+    }
+  }
+}.value
